Rename utils.ts to cn.ts
It’s the only used function in there
