+++ src/components/ui/form.tsx
@@ -11,7 +11,7 @@ import {
 	useFormState,
 } from 'react-hook-form';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 import { Label } from '@/components/ui/label';
 
 const Form = FormProvider;
